// src/csvdataintegrationapp.client/src/app/api/models/employee-dto.ts
/* tslint:disable */
/* eslint-disable */
export interface EmployeeDto {
  address?: string | null;
  address2?: string | null;
  dateOfBirth?: string;
  emailHome?: string | null;
  forenames?: string | null;
  id?: string;
  mobile?: string | null;
  payrollNumber?: string | null;
  postcode?: string | null;
  startDate?: string;
  surname?: string | null;
  telephone?: string | null;
}